Extract ScoreToGoal, don't show scores and goals for indicators without both
