+++ app/javascript/src/components/List/Indicator.js
@@ -1,21 +1,10 @@
 import React from "react"
 import PropTypes from "prop-types"
-import { useSelector } from "react-redux"
-import {
-  getCurrentScoreForIndicator,
-  getTargetScoreForIndicator,
-} from "../../config/selectors"
 import IndicatorActionList from "./IndicatorActionList"
-import rightArrow from "./right-arrow.svg"
+import ScoreToGoal from "./ScoreToGoal"
 
 const Indicator = (props) => {
   const indicator = props.indicator
-  const currentScore = useSelector((state) =>
-    getCurrentScoreForIndicator(state, indicator)
-  )
-  const targetScore = useSelector((state) =>
-    getTargetScoreForIndicator(state, indicator)
-  )
 
   return (
     <div
@@ -23,19 +12,7 @@ const Indicator = (props) => {
       data-benchmark-indicator-display-abbrev={indicator.display_abbreviation}
     >
       <div className="row bg-light-gray px-2 header">
-        <div className="col-2 d-flex align-items-center justify-content-center">
-          <span
-            className={`badge badge-pill badge-primary align-middle badge-rounded-circle color-value-${currentScore} mx-1 px-2`}
-          >
-            <span>{currentScore}</span>
-          </span>
-          <img className="mx-1" src={rightArrow} />
-          <span
-            className={`badge badge-pill color-value-${targetScore}  align-middle badge-rounded-circle mx-1 px-2`}
-          >
-            <span>{targetScore}</span>
-          </span>
-        </div>
+        <ScoreToGoal indicator={indicator} />
         <div className="col-10">
           <b>Benchmark {indicator.display_abbreviation}:</b>
           &nbsp;
